Fixes problem where users cannot view an internship record if the student's 'level' field is empty or invalid.
We should still be able to show the rest of the internship data, not just quit completely because that field is missing/invalid. An exception was being thrown in WebServiceDataProvider when sanity checking was done on the 'level' field. This commit removes that exception and instead sets the value of the field to null. Other changes:
* Add a warning to InternshipView to alert the user to the missing field.
* Add sanity checking to AddInternship command to detect the missing level field and show a nice error message.
* Refactor some of the error handling and redirecting logic in AddInternship so that the redirectToForm method is more generalized.